requirements.txt
asgiref==3.5.2
Django==3.1.6
django-environ==0.9.0
djangorestframework==3.12.2
pytz==2022.1
sqlparse==0.4.2
